+++ media/smart-card/src/hooks.ts
@@ -1,1 +1,2 @@
 export { useSmartLinkActions } from './state/hooks-external/useSmartLinkActions';
+export { useSmartLinkReload } from './state/hooks-external/useSmartLinkReload';